+++ tests/e2e/oauth_source.py
@@ -1,4 +1,4 @@
-"""Shared tooling for OAuth source e2e tests"""
+"""Shared tooling for source-into-application e2e tests"""
 
 from pathlib import Path
 from time import sleep
@@ -16,10 +16,24 @@ from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
 from authentik.blueprints.tests import apply_blueprint, reconcile_app
-from authentik.core.models import SourceUserMatchingModes, User
-from authentik.core.tests.utils import create_test_user
+from authentik.common.oauth.constants import (
+    SCOPE_OFFLINE_ACCESS,
+    SCOPE_OPENID,
+    SCOPE_OPENID_EMAIL,
+    SCOPE_OPENID_PROFILE,
+)
+from authentik.core.models import Application, SourceUserMatchingModes, User
+from authentik.core.tests.utils import create_test_cert, create_test_user
 from authentik.flows.models import Flow
-from authentik.lib.generators import generate_id
+from authentik.lib.generators import generate_id, generate_key
+from authentik.providers.oauth2.models import (
+    ClientType,
+    GrantType,
+    OAuth2Provider,
+    RedirectURI,
+    RedirectURIMatchingMode,
+    ScopeMapping,
+)
 from authentik.sources.oauth.models import OAuthSource
 from authentik.stages.identification.models import IdentificationStage
 from tests.decorators import SHADOW_ROOT_RETRIES, retry
@@ -27,6 +41,29 @@ from tests.decorators import SHADOW_ROOT_RETRIES, retry
 APP_URL = "http://localhost:9009"
 # A non-root app URL, to check the original destination survives the source login
 DEEP_LINK_URL = f"{APP_URL}/deep/link"
+AUTHORIZATION_FLOW = "default-provider-authorization-implicit-consent"
+REDIRECT_URI = f"{APP_URL}/auth/callback"
+
+
+def source_app_test(func):
+    """Blueprints and retry every source-into-application test needs"""
+    for deco in (
+        reconcile_app("authentik_crypto"),
+        apply_blueprint("system/providers-oauth2.yaml", "system/providers-saml.yaml"),
+        apply_blueprint(f"default/flow-{AUTHORIZATION_FLOW}.yaml"),
+        apply_blueprint(
+            "default/flow-default-source-authentication.yaml",
+            "default/flow-default-source-enrollment.yaml",
+            "default/flow-default-source-pre-authentication.yaml",
+        ),
+        apply_blueprint(
+            "default/flow-default-authentication-flow.yaml",
+            "default/flow-default-invalidation-flow.yaml",
+        ),
+        retry(),
+    ):
+        func = deco(func)
+    return func
 
 
 def initial_uri(entry: str) -> str:
@@ -46,28 +83,104 @@ class FlowStageMixin:
         """
         for attempt in range(SHADOW_ROOT_RETRIES):
             try:
-                flow_executor = self.get_shadow_root("ak-flow-executor")
-                return self.get_shadow_root(stage, flow_executor)
+                return self.get_shadow_root(stage, self.get_shadow_root("ak-flow-executor"))
             except DetachedShadowRootException, StaleElementReferenceException:
                 if attempt == SHADOW_ROOT_RETRIES - 1:
                     raise
-                self.logger.debug("Flow stage went stale, retrying", stage=stage, attempt=attempt)
                 sleep(1)
         return None
+
+    def fill_prompt(self, field: str, value: str):
+        """Fill in a single field of a prompt stage"""
+        element = self.get_stage_shadow_root("ak-stage-prompt").find_element(
+            By.CSS_SELECTOR, f"input[name={field}]"
+        )
+        element.click()
+        element.send_keys(value)
+        element.send_keys(Keys.ENTER)
+
+
+class OIDCAppMixin(FlowStageMixin):
+    """An OIDC application, served by the oidc-test-client container"""
+
+    # The client sends prompt=consent for offline_access, so a consent stage is shown
+    # even with an implicit-consent authorization flow
+    scopes = [SCOPE_OPENID, SCOPE_OFFLINE_ACCESS, SCOPE_OPENID_PROFILE, SCOPE_OPENID_EMAIL]
+
+    def setUp(self):
+        self.client_id = generate_id()
+        self.app_client_secret = generate_key()
+        self.application_slug = generate_id()
+        super().setUp()
+
+    def extra_scope_mappings(self) -> list[ScopeMapping]:
+        """Scope mappings on top of the defaults"""
+        return []
+
+    def setup_app(self):
+        provider = OAuth2Provider.objects.create(
+            name=self.application_slug,
+            client_type=ClientType.CONFIDENTIAL,
+            client_id=self.client_id,
+            client_secret=self.app_client_secret,
+            signing_key=create_test_cert(),
+            redirect_uris=[RedirectURI(RedirectURIMatchingMode.STRICT, REDIRECT_URI)],
+            authorization_flow=Flow.objects.get(slug=AUTHORIZATION_FLOW),
+            grant_types=[GrantType.AUTHORIZATION_CODE, GrantType.REFRESH_TOKEN],
+        )
+        provider.property_mappings.set(
+            [
+                *ScopeMapping.objects.filter(scope_name__in=self.scopes),
+                *self.extra_scope_mappings(),
+            ]
+        )
+        Application.objects.create(
+            name=self.application_slug, slug=self.application_slug, provider=provider
+        )
+        # The application has to exist before the client fetches the discovery document
+        self.run_container(
+            image="ghcr.io/beryju/oidc-test-client:2.7.1",
+            ports={"9009": "9009"},
+            environment={
+                "OIDC_CLIENT_ID": self.client_id,
+                "OIDC_CLIENT_SECRET": self.app_client_secret,
+                "OIDC_PROVIDER": f"{self.live_server_url}/application/o/{self.application_slug}/",
+                "OIDC_SCOPES": ",".join(self.scopes),
+            },
+        )
+
+    def pass_consent(self):
+        """Click through the consent stage the client asks for"""
+        try:
+            self.wait.until(ec.url_contains(f"/if/flow/{AUTHORIZATION_FLOW}/"))
+        except TimeoutException:
+            self.fail(
+                "Expected the pending application authorization to resume, but ended up at "
+                f"{self.driver.current_url}"
+            )
+        self.get_stage_shadow_root("ak-stage-consent").find_element(
+            By.CSS_SELECTOR, "[type=submit]"
+        ).click()
+
+    # The client always lands on its redirect URI; the URL the login started at comes
+    # back in the token payload as InitialURL instead
+    def app_destination_url(self) -> str:
+        return REDIRECT_URI
+
+    def deep_link_destination_url(self) -> str:
+        return REDIRECT_URI
 
 
 class AppRedirectMixin:
     """Assert the browser lands back at an application after logging in"""
 
     def wait_for_app(self, destination: str):
-        """Wait until the browser has landed back at the expected application URL"""
         try:
             self.wait.until(lambda driver: driver.current_url.startswith(destination))
         except TimeoutException:
             self.fail(
-                "Expected to be redirected back to the application at "
-                f"{destination} after logging in via the source, "
-                f"but ended up at {self.driver.current_url}"
+                f"Expected to be redirected back to the application at {destination} "
+                f"after logging in via the source, but ended up at {self.driver.current_url}"
             )
 
 
@@ -89,8 +202,7 @@ class DexOAuthSourceMixin(FlowStageMixin):
             environment={
                 "AK_HOST": self.host,
                 "AK_REDIRECT_URL": self.url(
-                    "authentik_sources_oauth:oauth-client-callback",
-                    source_slug=self.slug,
+                    "authentik_sources_oauth:oauth-client-callback", source_slug=self.slug
                 ),
                 "AK_CLIENT_SECRET": self.client_secret,
             },
@@ -126,7 +238,6 @@ class DexOAuthSourceMixin(FlowStageMixin):
     def login_via_oauth_provider(self):
         """Perform login at the OAuth provider (Dex)"""
         self.wait.until(ec.presence_of_element_located((By.ID, "login")))
-
         initial_provider_url = self.driver.current_url
 
         self.driver.find_element(By.ID, "login").send_keys("admin@example.com")
@@ -134,16 +245,8 @@ class DexOAuthSourceMixin(FlowStageMixin):
         self.driver.find_element(By.ID, "password").send_keys(Keys.ENTER)
 
         self.wait.until(ec.presence_of_element_located((By.CSS_SELECTOR, "button[type=submit]")))
-
         self.driver.find_element(By.CSS_SELECTOR, "button[type=submit]").click()
-
         self.wait.until(ec.url_changes(initial_provider_url))
-
-        self.assertNotEqual(
-            initial_provider_url,
-            self.driver.current_url,
-            "Expected to be redirected after login at OAuth provider",
-        )
 
     @retry(is_test_case=False)
     def click_source_button(self):
@@ -156,41 +259,13 @@ class DexOAuthSourceMixin(FlowStageMixin):
         )
         identification_stage.find_element(By.CSS_SELECTOR, selector).click()
 
-    def enroll_username(self, username: str):
-        """Fill in the username asked for by the source enrollment prompt"""
-        prompt_stage = self.get_stage_shadow_root("ak-stage-prompt")
-
-        prompt_stage.find_element(By.CSS_SELECTOR, "input[name=username]").click()
-        prompt_stage.find_element(By.CSS_SELECTOR, "input[name=username]").send_keys(username)
-        prompt_stage.find_element(By.CSS_SELECTOR, "input[name=username]").send_keys(Keys.ENTER)
-
 
 class SourceAppRedirectMixin(AppRedirectMixin, DexOAuthSourceMixin):
     """Log in at an application via an OAuth source, and assert we land back at the app.
 
-    Subclasses provide the application (`setup_app`), its claim assertions
-    (`assert_app_login`) and the URLs it is expected to be reached at; the tests
-    below are identical for every provider type.
+    Subclasses provide the application (`setup_app`, `*_destination_url`, `pass_consent`)
+    and its claim assertions (`assert_app_login`).
     """
-
-    def setup_app(self) -> None:
-        """Create the provider + application, and start the container acting as the app"""
-        raise NotImplementedError
-
-    def assert_app_login(self, user: User, entry_uri: str) -> None:
-        """Assert the app logged `user` in, having been entered at `entry_uri`"""
-        raise NotImplementedError
-
-    def app_destination_url(self) -> str:
-        """The URL the application is expected to be reached at once login is done"""
-        raise NotImplementedError
-
-    def deep_link_destination_url(self) -> str:
-        """Where a login started at `DEEP_LINK_URL` is expected to land"""
-        raise NotImplementedError
-
-    def pass_consent(self):
-        """Click through the consent stage, for applications that trigger one"""
 
     def restart_at_app(self, entry: str):
         """Start a fresh login at the app, discarding any session it kept for itself.
@@ -224,50 +299,36 @@ class SourceAppRedirectMixin(AppRedirectMixin, DexOAuthSourceMixin):
         self.driver.get(entry)
         self.click_source_button()
         self.login_via_oauth_provider()
-
         # At this point we've been redirected back and we're asked for the username
-        self.enroll_username("foo")
+        self.fill_prompt("username", "foo")
 
         self.pass_consent()
         self.wait_for_app(destination)
-        # `name` comes from dex's static user, `username` from the enrollment prompt
-        self.assert_app_login(
-            User(username="foo", name="admin", email="admin@example.com"), initial_uri(entry)
-        )
+        self.assert_app_login(self.enrolled_user(), initial_uri(entry))
 
-    @retry()
-    @apply_blueprint(
-        "default/flow-default-authentication-flow.yaml",
-        "default/flow-default-invalidation-flow.yaml",
-    )
-    @apply_blueprint(
-        "default/flow-default-source-authentication.yaml",
-        "default/flow-default-source-enrollment.yaml",
-        "default/flow-default-source-pre-authentication.yaml",
-    )
-    @apply_blueprint("default/flow-default-provider-authorization-implicit-consent.yaml")
-    @apply_blueprint("system/providers-oauth2.yaml", "system/providers-saml.yaml")
-    @reconcile_app("authentik_crypto")
+    def enrolled_user(self) -> User:
+        """`name` comes from dex's static user, `username` from the enrollment prompt"""
+        return User(username="foo", name="admin", email="admin@example.com")
+
+    @source_app_test
     def test_source_auth(self):
         """test app login via OAuth source (existing user, linked by email)"""
         self.source_auth(APP_URL, self.app_destination_url())
 
-    @retry()
-    @apply_blueprint(
-        "default/flow-default-authentication-flow.yaml",
-        "default/flow-default-invalidation-flow.yaml",
-    )
-    @apply_blueprint(
-        "default/flow-default-source-authentication.yaml",
-        "default/flow-default-source-enrollment.yaml",
-        "default/flow-default-source-pre-authentication.yaml",
-    )
-    @apply_blueprint("default/flow-default-provider-authorization-implicit-consent.yaml")
-    @apply_blueprint("system/providers-oauth2.yaml", "system/providers-saml.yaml")
-    @reconcile_app("authentik_crypto")
+    @source_app_test
     def test_source_enroll(self):
         """test app login via OAuth source (new user, enrolled)"""
         self.source_enroll(APP_URL, self.app_destination_url())
+
+    @source_app_test
+    def test_source_auth_deep_link(self):
+        """test app login via OAuth source, started at a deep link into the app"""
+        self.source_auth(DEEP_LINK_URL, self.deep_link_destination_url())
+
+    @source_app_test
+    def test_source_enroll_deep_link(self):
+        """test app enrollment via OAuth source, started at a deep link into the app"""
+        self.source_enroll(DEEP_LINK_URL, self.deep_link_destination_url())
 
     @retry()
     def test_source_enroll_auth(self):
@@ -285,41 +346,4 @@ class SourceAppRedirectMixin(AppRedirectMixin, DexOAuthSourceMixin):
 
         self.pass_consent()
         self.wait_for_app(self.app_destination_url())
-        # `name` comes from dex's static user, `username` from the enrollment prompt
-        self.assert_app_login(
-            User(username="foo", name="admin", email="admin@example.com"), initial_uri(APP_URL)
-        )
-
-    @retry()
-    @apply_blueprint(
-        "default/flow-default-authentication-flow.yaml",
-        "default/flow-default-invalidation-flow.yaml",
-    )
-    @apply_blueprint(
-        "default/flow-default-source-authentication.yaml",
-        "default/flow-default-source-enrollment.yaml",
-        "default/flow-default-source-pre-authentication.yaml",
-    )
-    @apply_blueprint("default/flow-default-provider-authorization-implicit-consent.yaml")
-    @apply_blueprint("system/providers-oauth2.yaml", "system/providers-saml.yaml")
-    @reconcile_app("authentik_crypto")
-    def test_source_auth_deep_link(self):
-        """test app login via OAuth source, started at a deep link into the app"""
-        self.source_auth(DEEP_LINK_URL, self.deep_link_destination_url())
-
-    @retry()
-    @apply_blueprint(
-        "default/flow-default-authentication-flow.yaml",
-        "default/flow-default-invalidation-flow.yaml",
-    )
-    @apply_blueprint(
-        "default/flow-default-source-authentication.yaml",
-        "default/flow-default-source-enrollment.yaml",
-        "default/flow-default-source-pre-authentication.yaml",
-    )
-    @apply_blueprint("default/flow-default-provider-authorization-implicit-consent.yaml")
-    @apply_blueprint("system/providers-oauth2.yaml", "system/providers-saml.yaml")
-    @reconcile_app("authentik_crypto")
-    def test_source_enroll_deep_link(self):
-        """test app enrollment via OAuth source, started at a deep link into the app"""
-        self.source_enroll(DEEP_LINK_URL, self.deep_link_destination_url())
+        self.assert_app_login(self.enrolled_user(), initial_uri(APP_URL))
